morning_brief/requirements.txt
requests>=2.31.0
feedparser>=6.0.10
schedule>=1.2.1
reportlab>=4.2.0
openai>=1.30.0
python-dotenv>=1.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
yfinance>=0.2.38
fredapi>=0.5.1
finnhub-python>=2.4.19
pandas>=2.2.2
